+++ src-tauri/src/service/local/voice_clone.rs
@@ -4,17 +4,20 @@ use anyhow::{bail, Context};
 use sea_orm::{
     ActiveModelTrait, ActiveValue::NotSet, ActiveValue::Set, EntityTrait, TransactionTrait,
 };
+use tokio::sync::watch;
 
 use crate::{
     common::{
         local_paths::{ensure_child_dir, resolve_task_path, serialize_task_path},
         task_paths::ensure_task_sample_dir,
     },
+    config::BaseModel,
     service::{
         local::entity::{
             task_history as task_history_entity, voice_clone_task as voice_clone_task_entity,
         },
         models::{CreateVoiceCloneTaskPayload, HistoryTaskType, TaskStatus, VoiceCloneTaskResult},
+        pipeline::{resolve_model_task_pipeline, VoiceClonePipelineRequest},
         LocalService,
     },
     utils::time::now_string,
@@ -164,5 +167,38 @@ impl LocalService {
             status: TaskStatus::Pending,
             output_file_path: serialized_output_path,
         })
+    }
+
+    pub(crate) fn start_voice_clone_inference(
+        &self,
+        base_model: BaseModel,
+        task_id: i64,
+    ) -> Result<()> {
+        let service = self.clone();
+        let pipeline = resolve_model_task_pipeline(&base_model)?;
+        let (cancel_tx, cancel_rx_guard) = watch::channel(false);
+        self.register_active_task_control(
+            task_id,
+            HistoryTaskType::VoiceClone,
+            cancel_tx,
+            cancel_rx_guard,
+        );
+
+        tauri::async_runtime::spawn(async move {
+            let result = pipeline
+                .run_voice_clone_pipeline(
+                    base_model.to_string(),
+                    &service,
+                    VoiceClonePipelineRequest { task_id },
+                )
+                .await;
+            service.unregister_active_task_control(task_id);
+
+            if let Err(err) = result {
+                tracing::error!(error = %err, "local voice clone pipeline failed");
+            }
+        });
+
+        Ok(())
     }
 }
